requirements.txt
Flask
Flask-SQLAlchemy
Werkzeug
SQLAlchemy
APScheduler
python-dotenv
Flask-Migrate